Added new settings
Add an option to make LSCWP's cache crawler feature run within a certain range of time only
Add an option to whitelist certain HTML comments from being minified; this is needed for some plugins, as suggested here: https://wordpress.org/support/topic/feature-request-option-to-disable-removing-comments-when-minifing-html/
Add an option to keep guest mode pages and only purge normal pages when using LSCWP's "Purge All" function.

+++ src/purge.cls.php
@@ -192,7 +192,22 @@ class Purge extends Base
 		// 	self::debug( 'CLI request, queue stored: ' . $curr_built );
 		// }
 		// else {
-		$this->_purge_all_lscache(true);
+
+		$purge_guest = $this->conf(BASE::O_PURGE_GUEST_PAGES);
+		if ($purge_guest) {
+			$this->_purge_all_lscache(true);
+		}
+		//  else {
+		// 	$this->_add(Tag::TYPE_LOCALRES);
+		// 	$this->_add(Tag::TYPE_HTTP);
+		// 	$this->_add(Tag::TYPE_MIN);
+		// 	// $this->_add(Tag::TYPE_FRONTPAGE);
+		// 	// if (LITESPEED_SERVER_TYPE !== 'LITESPEED_SERVER_OLS') {
+		// 	// 	$this->_add_private(Tag::TYPE_FRONTPAGE);
+		// 	// }
+		// 	// $this->_add(Tag::TYPE_PAGES);
+		// }
+
 		$this->_purge_all_cssjs(true);
 		$this->_purge_all_localres(true);
 		// $this->_purge_all_ccss( true );
